+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.5.0
 
-* Implemented new cookie manager for setting cookies and providing initial cookies.
+* Implements new cookie manager for setting cookies and providing initial cookies.
 
 ## 2.4.0